pyproject.toml
[tool.poetry]
name = "amalthea"
version = "0.6.0"
description = "Operator for Jupyter servers in K8s"
authors = ["SDSC"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
kopf = "^1.36.0"
pyyaml = "^6.0"
kubernetes = "24.2.0"
jsonpatch = "^1.32"
json-merge-patch = "^0.2"
expiringdict = "^1.2.2"
jinja2 = "^3.1.2"
pytz = "^2022.6"
prometheus-client = "^0.15.0"
dataconf = "^2.1.3"
boto3 = "^1.26.31"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"
pytest = "^7.2.0"
pylint = "^2.15.8"
pytest-mock = "^3.10.0"
chartpress = "^2.1.0"
flake8-pyproject = "^1.2.2"
isort = "^5.11.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = ["-v"]
testpaths = [
    "controller",
    "tests",
    "utils",
]
log_cli = true

[tool.flake8]
ignore = ["E203", "W503"]
max-line-length = 100
count = true
per-file-ignores = [
    '__init__.py:F401',
    'kopf_entrypoint.py:F401',
]

[tool.black]
line-length = 100

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 100
